// src/generated/openapi/models/MessageDTO.ts
// @ts-nocheck - Auto-generated file, TypeScript strict checks disabled intentionally
/* tslint:disable */
/* eslint-disable */
/**
 * VIPLEV
 * This is an API specification
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MessageDTO
 */
export interface MessageDTO {
    /**
     * ID of the benchmark
     * @type {string}
     * @memberof MessageDTO
     */
    benchmarkId?: string;
    /**
     * ID of the run
     * @type {string}
     * @memberof MessageDTO
     */
    runId?: string;
    /**
     * Type of message for the agent
     * @type {MessageDTOMessageTypeEnum}
     * @memberof MessageDTO
     */
    messageType?: MessageDTOMessageTypeEnum;
}


/**
 * @export
 */
export const MessageDTOMessageTypeEnum = {
    PendingStart: 'PENDING_START',
    PendingStop: 'PENDING_STOP'
} as const;
export type MessageDTOMessageTypeEnum = typeof MessageDTOMessageTypeEnum[keyof typeof MessageDTOMessageTypeEnum];


/**
 * Check if a given object implements the MessageDTO interface.
 */
export function instanceOfMessageDTO(value: object): value is MessageDTO {
    return true;
}

export function MessageDTOFromJSON(json: any): MessageDTO {
    return MessageDTOFromJSONTyped(json, false);
}

export function MessageDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): MessageDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'benchmarkId': json['benchmarkId'] == null ? undefined : json['benchmarkId'],
        'runId': json['runId'] == null ? undefined : json['runId'],
        'messageType': json['messageType'] == null ? undefined : json['messageType'],
    };
}

export function MessageDTOToJSON(json: any): MessageDTO {
    return MessageDTOToJSONTyped(json, false);
}

export function MessageDTOToJSONTyped(value?: MessageDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'benchmarkId': value['benchmarkId'],
        'runId': value['runId'],
        'messageType': value['messageType'],
    };
}